feat(supplier): add flexible filter endpoint
- Add filterAllByValue method in controller and service
- Support filtering by numeric or string input
- Add custom multi-field query in repository
- Fix message typo in HandlerExceptionHandler
- Fix regex property name in RegexPatterns

+++ backend/administrative-system-backend/src/main/java/com/gg/administrative_system_backend/contract/ContractService.java
@@ -40,7 +40,7 @@ public class ContractService {
             throw new ValueRequiredException(ContractMessage.VALUE_REQUIRED.getMessage());
         }
         if(value.matches(RegexPatterns.BIG_DECIMAL)){
-            if(value.matches(RegexPatterns.FLOAT)){
+            if(value.matches(RegexPatterns.LONG)){
                 return contractRepository.findById(Long.parseLong(value)).map(List::of).orElse(List.of());
             } else {
                 return contractRepository.findByTotalExpenses(new BigDecimal(value.replace(",", "")));
