+++ backend/administrative-system-backend/src/main/java/com/gg/administrative_system_backend/util/RegexPatterns.java
@@ -2,5 +2,5 @@ package com.gg.administrative_system_backend.util;
 
 public class RegexPatterns {
     public static String BIG_DECIMAL = "[\\d,]+(\\.\\d+)?";
-    public static String FLOAT = "\\d+";
+    public static String LONG = "\\d+";
 }
